Updated files
These are the latest updates before I move to making the input an object/dictionary. These scripts have fixed the inverted and negative values in the output curves. That had to do with the "ghost" cell defined at each boundary. We should talk about that next time so we can clarify what a good assumption is for those.