Input is now read through functions
diffOpts is my simpler version of diffusionOpts. It reads in the information from input*.inp
nuclide.py reads in information from the files contained in the Materials folder (ex: U235.inp)
reader.sh does not read any information in anymore, but it does pass the name of the input file, since its job is to run multiple input files.