Latest updates
This update creates a slab that's half U-235 and half U-238. It is also a multigroup solver: it solves the problem for each input file as many times as there are groups specified in that input file. For a given energy group, it calls a new material (ex: group 1 calls U-235(1) which has different cross sections than U235(2)). 
The output looks good to me because it reflects the problem accurately. The half of the slab with U-235 behaves differently than the other half with U-238, and the different curves represent different energy groups.