Update 7/12/2016
With this update, you no longer run ./reader.sh, but you run python diffusion.py. The code calls a class which grabs the input file names. The code also now calls a class at the end which plots the results. This update solves the problem for a slab of a single nuclide but any number of energy groups (as long as there’s a material file in the Materials folder with that energy group number) and plots them all in one graph. The plotting happens in a separate class. 
I still want to:
-make the material files more like a dictionary 
-make the user define the slab composition from the input file
-get the macro xs from number density*microxs