Latest updates & bug
Some of the curves weren’t coming down to 0 at the ends. If the scattering cross section was too high, say 0.05, it wouldn’t work, but with lower numbers, it was fixed. However, I think the calculation should work regardless of the values, which raises a red flag. What do you think?
The source works really well as long as it’s far from the edges. See the figures in the output folder. If the source is close to the edge, I get negative fluxes. I tried a few approaches to narrow down the bug, but I think at this point we need to talk through it together. It might be in my boundary conditions, which I still don’t fully understand. 
I still have not had time to work on: input file slab composition, materials