Updates
I found a typo in the boundary condition, a little “/” has slid itself in a random space. Fixing this fixed the problems I was having. The ends of the curves all hit 0 no matter what the scattering cross sections were, and the sources close to the boundaries made much more sense. 
Then, I added the “forward-looking” diffusion coefficient in the boundary conditions, and it all got worse.