updated files: diffusion.py, run.sh
This update features the arbitrary source. It can be modified to have as many arbitrary sources as you want in a given run.
To run this version, run ./run.sh and assign n in that script, the number of times you want to run the problem.
For now, I have commented out the plotting portion, and I write the source location to the file “sources” in the output folder. 
You can see that with many runs, the source tends to end up in the higher fission XS group, which is what I intended. I realize that we are going to want to store the solution somewhere, but I am not yet sure what “convergence” really looks like or how we find it, so I held off on trying to keep the solution anywhere.