Updates
I assigned number densities for 3 materials (fuel, moderator, poison) to each bin. So the matrix numDensity has 3 columns, one for each material, and numBin rows, for each bin. Besides the fuel (column 1) the other two columns (moderator and poison) are ignored for now because they don’t play a role in the code yet. For now, the number densities are 1 because when I put in reasonable numbers such as 2.3e22, the curves were flat. 
Then, the macroscopic cross sections are calculated by multiplying the microscopic cross section (from the material input files) with the number density for that bin. 
Starting at line 84: When solving the problem multiple times, the number densities are cut in half. The cross sections are then re-calculated. The exception is the scattering kernel which gets re-calculated in the constructA subroutine. 
I would find it helpful if we went through this change together to make sure I was thorough. We should also see about actual number density values.